Remote: extract the v4 transfer bookkeeping and the legacy consent queue
Two cohesive clusters leave `remote_command_router.dart`, each as a real
object that owns its own state and takes what it needs through constructor
callbacks.

`lib/services/remote_control/remote_transfer_bookkeeping.dart` —
`RemoteTransferBookkeeper`. The open v4 transaction (`_openRemoteTransfer`'s
bookkeeping half, `_recordRemoteTransferCommand`, `_remoteTransferItemPayload`,
`_activeRemoteTransferContainsExpected`, `_activeRemoteTransferMatches`,
`_clearActiveRemoteTransfer`, `_remoteBatchCommands`) plus the retained
outcomes and the reports that fill them (`_remoteTransferOutcomes`,
`_remoteTransferOutcomeLifetime`, `_reportProfileGraphResult`,
`_runBestEffortProfileGraphResult`, `_reportProfileGraphResultBestEffort`,
`_addonTransferRequestId`, `_reportAddonTransferResultBestEffort`,
`_reportRemoteTransferResultBestEffort`,
`_reportCompleteTransferResultBestEffort`). It gets the peer-key derivation
through a callback; the router keeps the parts of `_openRemoteTransfer` that
touch its own staging buffer, batcher and in-flight set.

`lib/services/remote_control/remote_legacy_consent.dart` —
`RemoteLegacyConsentQueue` and `LegacyConsentItem` (was `_LegacyItem`). The
plaintext (v1) buffer, its 60s expiry, the approval window and the approved
replay (`_legacyApprovedFor`, `_enqueueLegacy`,
`_maybeShowLegacyConsentDialog`, `_allowLegacy`, `_denyLegacy`). The dialog is
the only part that needs a `BuildContext`, so it stays in the router as
`_presentLegacyConsent` / `_dismissLegacyConsent`; the queue asks for it,
is told the answer through `onConsentAnswer`, and gets dispatch, the snackbar
sink, the batcher and the activity stamp as callbacks.

Preserved quirks: the outcome cache still answers a retried `complete` before
any manifest check, so a lost acknowledgement cannot apply a payload twice; a
refused `complete` is still reported, on the addon channel when the body is a
bare request id and not at all when that id exceeds 128 characters; results are
still sent three times and every failure is still swallowed; a wrapped item is
still meaningful only inside its own open transaction while a raw body stays
valid for non-transactional senders; consent still belongs to ONE datagram
source, an Allow behind an empty buffer still grants nothing, and the parked
`complete` is still replayed last.

Neither new file references a router private, and neither imports Flutter
widgets, `lib/widgets` or `lib/screens` — layering stays at 53 violations with
no new rows. One debug seam is kept: `debugRunBestEffortProfileGraphResult`
now delegates to the bookkeeper (used by `remote_chunked_send_test.dart`). A
stale doc comment about the profile-graph import, which had drifted onto
`_reportProfileGraphResult`, moves down onto `_handleProfileGraphConfig` where
it belongs.

Pinned by `test/remote_transfer_bookkeeping_origin_pin_test.dart`, unchanged
from the previous commit and still passing. Mutation check repeated after the
move: dropping the outcome write and dropping the different-peer guard each
fail their test in the new files. Host: 4300 -> 3887 lines. No forwarders.
